userauthentication

+++ controllers/user.js
@@ -1,5 +1,5 @@
 const bcrypt = require('bcrypt');
-
+const jwt = require('jsonwebtoken')
 
 
 const signup_details = require('../models/userInfo.js');
@@ -9,6 +9,18 @@ const signup_table = signup_details.Signup_details;
 
 const expense_table = require('../models/expenses.js');
 
+
+
+
+
+
+
+
+
+function generateAccessToken(id){
+    const token = jwt.sign({userId:id},'myToken');
+    return token;
+}
 
 exports.signUp = (req,res,next)=>{
     const body = req.body;
@@ -55,18 +67,23 @@ exports.login = (req,res,next) =>{
     const body = req.body;
     console.log(req.body.email)
 
+    
+
     signup_table.findAll({where:{email:req.body.email}}).then(data=>{
         
         if(data[0]){
             bcrypt.compare(req.body.password,data[0].dataValues.password,(err,response)=>{
-               console.log(response);
+              // console.log(response);
+               const id = data[0].dataValues.id;
+               const token = generateAccessToken(id);
+               console.log(token);
 
                if(err){
                 res.status(500).send('something went wrong')
                }
 
              if(response===true){
-                 res.send('User login sucessful')
+                res.json({id:token});
              }
              else if(response===false){
                  res.status(401).send('User not authorised')
@@ -91,12 +108,17 @@ exports.login = (req,res,next) =>{
 
 exports.postExpenses=(req,res,next)=>{
 
+
     const incomingData=req.body;
+    const id = req.userId;
+    console.log('this is your id:',id);
+
 
     expense_table.create({
        amount: incomingData.amount,
        category: incomingData.category,
-       description: incomingData.description
+       description: incomingData.description,
+       signupId:id,
     })
     .then(serverData=>{
         res.json({data:serverData.dataValues})
@@ -107,10 +129,11 @@ exports.postExpenses=(req,res,next)=>{
 
 exports.getExpenses=(req,res,next)=>{
 
-    expense_table.findAll()
+   // console.log("user id-",req.userId);
+   const id = req.userId;
+    expense_table.findAll({where:{signupId:id}})
     .then((result)=>{
         const pureResult = result.map(key=>key.dataValues)
-       
         res.json(pureResult);
 
     })
@@ -124,8 +147,8 @@ exports.getExpenses=(req,res,next)=>{
 
 exports.deleteExpense=(req,res,next)=>{
     const id =req.params.id;
-
-    expense_table.destroy({where:{id:id}})
+    const userid = req.userId;
+    expense_table.destroy({where:{id:id,signupId:userid}})
     .then((result)=>{
           res.json({message:'User deleted'});
       })
